Load rosetta file relative location
